+++ e2e/src/fixtures/metamask.fixture.ts
@@ -1,12 +1,12 @@
 import { loadEnvConfig } from '@config/env.js'
 import { EXTENSION_TIMEOUTS } from '@constants/timeouts.js'
-import { VIEWPORT } from '@constants/viewport.js'
 import { MetaMaskPage } from '@pages/metamask/metamask.page.js'
-import { chromium, test as base } from '@playwright/test'
+import { test as base } from '@playwright/test'
 import fs from 'node:fs'
-import os from 'node:os'
-import path from 'node:path'
 
+import { launchExtensionContext } from './extension-context.js'
+
+import type { LaunchExtensionOptions } from './extension-context.js'
 import type { BrowserContext, Page } from '@playwright/test'
 
 export interface MetaMaskFixtures {
@@ -16,17 +16,13 @@ export interface MetaMaskFixtures {
   hubPage: Page
 }
 
-export interface LaunchMetaMaskOptions {
-  /** Called with the extension path BEFORE the browser reads the extension files. */
-  beforeLaunch?: (extensionPath: string) => void | Promise<void>
-  /** Called after the browser context is closed. */
-  afterClose?: () => void | Promise<void>
-  /** Additional Chrome flags appended to the default set. */
-  extraChromeArgs?: string[]
-}
+export type LaunchMetaMaskOptions = LaunchExtensionOptions
 
 /**
  * Launch a persistent Chromium context with MetaMask loaded.
+ *
+ * Thin wrapper over {@link launchExtensionContext} that resolves the MetaMask
+ * extension path from env and gives a MetaMask-specific "not found" hint.
  */
 export async function launchMetaMaskContext(
   use: (context: BrowserContext) => Promise<void>,
@@ -41,34 +37,10 @@ export async function launchMetaMaskContext(
     )
   }
 
-  await options.beforeLaunch?.(extensionPath)
-
-  const profileDir = fs.mkdtempSync(path.join(os.tmpdir(), 'pw-metamask-'))
-
-  const launchStart = Date.now()
-  const context = await chromium.launchPersistentContext(profileDir, {
-    headless: false,
-    args: [
-      `--disable-extensions-except=${extensionPath}`,
-      `--load-extension=${extensionPath}`,
-      '--no-first-run',
-      '--disable-default-apps',
-      ...(options.extraChromeArgs ?? []),
-    ],
-    viewport: { width: VIEWPORT.WIDTH, height: VIEWPORT.HEIGHT },
+  await launchExtensionContext(extensionPath, use, {
+    profilePrefix: 'pw-metamask-',
+    ...options,
   })
-  console.log(
-    `[metamask-fixture] Browser launched in ${Date.now() - launchStart}ms`,
-  )
-
-  await use(context)
-
-  try {
-    await context.close()
-  } finally {
-    fs.rmSync(profileDir, { recursive: true, force: true })
-    await options.afterClose?.()
-  }
 }
 
 export const test = base.extend<MetaMaskFixtures>({
